Civilization Preparation
Made updates to give us a framework towards working on the civilization and resource systems. New script to manage the growth of each of those.

Resources grow when the correct sensor states are reached for sufficient time. They decay when conditions are not met, but cannot regress stages.

Civilization script tracks buildings (so the models can be updated as progression) and peeps (little capsules that will be our people). Both have been made Prefabs, so we may edit them easier and add behaviors. Controller script is very barebones at the moment.